Add title attribute to iframe
It's disputable if single inline iframes should have a title
attribute for improving a11y, yet Lighthouse complains.

Please cmp. https://web.dev/frame-title/ or
https://dequeuniversity.com/rules/axe/3.3/frame-title

This pull request adds the title attribute using the content title
to the embedding iframe and to the embed code that can be
retrieved. If accepted, this should also be done for the Drupal
plugin and the moodle plugin.